fix(acp): dedupe utf8PrefixWithin into a shared internal module
Three byte-identical private copies of the UTF-8-boundary prefix helper
were introduced across live-record.ts, the ACP reducer, and
tool-presentation-fold.ts, replacing the single copy deleted earlier from
streaming-output.ts. The continuation-byte back-off logic is subtle
enough that verbatim triplication is real drift risk. New
presentation/utf8.ts, imported by all three; no pi-agent-core public API
change -- a coding-agent-internal module, not a new shared-package
export.
